pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "vpn-manager"
dynamic = ["version"]
description = "Advanced VPN manager with kill switch and IP rotation"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "VPN Manager Team", email = "contact@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "Topic :: Security",
    "Topic :: Internet",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: POSIX :: Linux",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0,<3.0.0",
    "pyyaml>=6.0,<7.0",
    "rich>=13.0.0,<14.0.0",
    "click>=8.1.0,<9.0.0",
    "dnspython>=2.4.0,<3.0.0",
    "textual>=0.47.0,<1.0.0",
    "urllib3>=2.0.0,<3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "pytest-timeout>=2.1.0,<3.0.0",
    "pytest-mock>=3.11.0,<4.0.0",
    "black>=23.7.0,<24.0.0",
    "flake8>=6.1.0,<7.0.0",
    "mypy>=1.5.0,<2.0.0",
    "pylint>=2.17.0,<3.0.0",
    "sphinx>=7.1.0,<8.0.0",
    "sphinx-rtd-theme>=1.3.0,<2.0.0",
    "textual-dev>=1.0.0,<2.0.0",
]

[project.scripts]
vpn-manager = "vpn_manager.cli.interface:main"

[tool.setuptools_scm]
write_to = "vpn_manager/_version.py"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=vpn_manager --cov-report=html --cov-report=term-missing"
